requirement.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.2.9
drf-spectacular==0.28.0
djoser==2.3.3